Disable blackbox macro expansion
