+++ test-workspace/build.sbt
@@ -3,3 +3,8 @@ inThisBuild(
     scalaVersion := "2.12.8"
   )
 )
+
+libraryDependencies ++= List(
+  "com.lihaoyi" %% "pprint" % "0.5.2",
+  "com.chuusai" %% "shapeless" % "2.3.3"
+)
